Fix d^2(I3)/dE^2 error and d(\sigma)/dp ommision in analytic Jacobian
Two negative signs missing in evaluation of d^2(I3)/dE^2. Also, assumption that the Jacobian matrix is perfectly symmetrical is not necessarily true due to the hydrostatic pressure formulation. Put in the evaluation of a hydrostatic portion of the elasticity tensor for Guccione and Mooney Rivlin and add this portion to the final columns of the Jacobian matrix.