Fixes for dataprojections
Distance vector sign was updated in previous commits by setting:
distanceVector(1:numberOfCoordinates)=interpolatedPoint%values(:,NO_PART_DERIV)-dataPointLocation
and by multiplying gradient and Hessian values by -1 compared with the original implementation. 

However, with the updated code, the gradient values are the same as what the old code produced, however the Hessian isn't. Fix was to revert to the original implementation for the gradient and Hessian and setting the distance vector as:
distanceVector(1:numberOfCoordinates)=dataPointLocation-interpolatedPoint%values(:,NO_PART_DERIV) 
This has been done for all cases of the projection (line, face, element).